The time line merge from KO/Pi introduceda nasty regression (more than 24 hours are shown in a day!). This is the first shot, which works in most cases, but still lots of things remain to be fixed.
CCBUG:108606

svn path=/trunk/KDE/kdepim/; revision=438286

+++ koagendaview.h
@@ -75,7 +75,7 @@ class TimeLabels : public QScrollView
 
   private:
     int mRows;
-    int mCellHeight;
+    double mCellHeight;
     int mMiniWidth;
     KOAgenda* mAgenda;
 
@@ -143,7 +143,7 @@ class KOAgendaView : public KOEventView
     KOAgendaView( Calendar *cal, QWidget *parent = 0, const char *name = 0 );
     virtual ~KOAgendaView();
 
-    
+
 
     /** Returns maximum number of days supported by the koagendaview */
     virtual int maxDatesHint();
@@ -210,10 +210,10 @@ class KOAgendaView : public KOEventView
 
     void zoomInHorizontally( const QDate& date=QDate() );
     void zoomOutHorizontally( const QDate& date=QDate() );
-    
+
     void zoomInVertically( );
     void zoomOutVertically( );
-    
+
     void zoomView( const int delta, const QPoint &pos,
       const Qt::Orientation orient=Qt::Horizontal );
   signals:
